feat: improve testnet portfolio and total balance (#104)
